refactor(global): restructured
Restructured core files
