+++ .gitignore
@@ -1,6 +1,6 @@
 node_modules/**
 .idea/**
 build/**
-.env
+.dev.env
 .dev.env
 coverage
