+++ package.json
@@ -6,7 +6,7 @@
     "build": "tsc",
     "watch-ts": "tsc -w",
     "watch-node": "nodemon build/index.js",
-    "watch": "concurrently -k -p \"[{name}]\" -n \"TypeScript,Node\" -c \"yellow.bold,green.bold\" \"npm run watch-ts\" \"npm run watch-node\"",
+    "watch": "concurrently -k -p \"[{name}]\" -n \"TypeScript,Node\" -c \"yellow.bold,green.bold\" \"npm run watch-ts\" \"tsc-alias -w\" \"npm run watch-node\"",
     "start": "NODE_ENV=production npm run serve",
     "serve": "node build/index.js",
     "bump": "standard-version --skip.tag",
@@ -34,6 +34,7 @@
     "nodemon": "^2.0.12",
     "standard-version": "^9.3.1",
     "ts-jest": "^27.0.3",
+    "tsc-alias": "^1.6.7",
     "typescript": "^4.3.5"
   },
   "config": {
