+++ src/common/command.ts
@@ -1,4 +1,4 @@
-import { CommandOpts } from "./types";
+import { CommandOpts } from "../types";
 
 export class Command {
   constructor(public options: CommandOpts) {}
