+++ src/core/command-recognizer/findCommandInMessage.ts
@@ -1,4 +1,4 @@
-import { Command } from "../command";
+import { Command } from "@common/command";
 
 type FindCommandInMessageResult =
   { founded: false, command: null } |
