+++ src/core/command-resolver/command-resolver.ts
@@ -1,11 +1,11 @@
 import { CommonUserstate } from "tmi.js";
-import { Command } from "./command";
-import { CommandResolverOpts } from "./types";
-import { client } from "./index";
-import { commandRecognizer } from "./utils/commandRecognizer";
-import { messageToUser } from "./utils/messageToUser";
-import { checkRoles } from "./utils/checkRoles";
-import { throttlingManager } from "./throttling-manager";
+import { Command } from "@common/command";
+import { messageToUser } from "@utils/messageToUser";
+import { checkRoles } from "@utils/checkRoles";
+import { CommandResolverOpts } from "../../types";
+import { client } from "../../index";
+import { commandRecognizer } from "../command-recognizer/command-recognizer";
+import { throttlingManager } from "../throttling-manager/throttling-manager";
 
 export const resolveCommands = (commands: Command[], options?: CommandResolverOpts) => {
   const defaultOptions: CommandResolverOpts = {
