+++ src/core/throttling-manager/throttling-manager.ts
@@ -1,4 +1,4 @@
-import { Command } from "./command";
+import { Command } from "@common/command";
 
 export class ThrottlingManager {
   stack: NewThrottlingStackItem[] = [];
