+++ src/types.d.ts
@@ -1,4 +1,4 @@
-import { Command } from "./command";
+import { Command } from "./common/command";
 
 export type Throttling = {
   usages: number;
