+++ tsconfig.json
@@ -68,6 +68,13 @@
 
     /* Advanced Options */
     "skipLibCheck": true,                           /* Skip type checking of declaration files. */
-    "forceConsistentCasingInFileNames": true        /* Disallow inconsistently-cased references to the same file. */
+    "forceConsistentCasingInFileNames": true        /* Disallow inconsistently-cased references to the same file. */,
+    "baseUrl": ".",
+    "paths": {
+      "@core/*": ["./src/core/*"],
+      "@commands/*": ["./src/commands/*"],
+      "@common/*": ["./src/common/*"],
+      "@utils/*": ["./src/utils/*"]
+    }
   }
 }
